Reorganize Common module
