+++ src/main/java/ch/njol/skript/classes/data/SkriptClasses.java
@@ -23,8 +23,6 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.skriptlang.skript.bukkit.base.types.ItemTypeClassInfo;
 import org.skriptlang.skript.bukkit.base.types.SlotClassInfo;
-import org.skriptlang.skript.common.types.QueueClassInfo;
-import org.skriptlang.skript.common.types.ScriptClassInfo;
 import org.skriptlang.skript.lang.properties.Property;
 import org.skriptlang.skript.lang.properties.handlers.ContainsHandler;
 import org.skriptlang.skript.lang.properties.handlers.TypedValueHandler;
@@ -483,8 +481,6 @@ public class SkriptClasses {
 				.serializer(new YggdrasilSerializer<GameruleValue>())
 		);
 
-		Classes.registerClass(new QueueClassInfo());
-
 		Classes.registerClass(new ClassInfo<>(Config.class, "config")
 			.user("configs?")
 			.name("Config")
@@ -566,8 +562,6 @@ public class SkriptClasses {
 						return String.class;
 					}
 				}));
-
-		Classes.registerClass(new ScriptClassInfo());
 
 		Classes.registerClass(new ClassInfo<>(Executable.class, "executable")
 			.user("executables?")
