+++ src/main/java/ch/njol/skript/expressions/ExprAmount.java
@@ -18,7 +18,7 @@ import ch.njol.util.Kleenean;
 import ch.njol.util.coll.CollectionUtils;
 import org.bukkit.event.Event;
 import org.jetbrains.annotations.Nullable;
-import org.skriptlang.skript.common.properties.expressions.PropExprAmount;
+import org.skriptlang.skript.common.properties.elements.expressions.PropExprAmount;
 
 /**
  * @deprecated This is being removed in favor of {@link PropExprAmount}
