+++ src/main/java/org/skriptlang/skript/common/elements/expressions/ExprRecursiveSize.java
@@ -1,10 +1,11 @@
-package org.skriptlang.skript.common.expressions;
+package org.skriptlang.skript.common.elements.expressions;
 
 import ch.njol.skript.Skript;
 import ch.njol.skript.doc.Description;
 import ch.njol.skript.doc.Example;
 import ch.njol.skript.doc.Name;
 import ch.njol.skript.doc.Since;
+import ch.njol.skript.expressions.base.PropertyExpression;
 import ch.njol.skript.lang.*;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.lang.util.SimpleExpression;
@@ -12,6 +13,8 @@ import ch.njol.skript.util.LiteralUtils;
 import ch.njol.util.Kleenean;
 import org.bukkit.event.Event;
 import org.jetbrains.annotations.Nullable;
+import org.skriptlang.skript.registration.SyntaxInfo;
+import org.skriptlang.skript.registration.SyntaxRegistry;
 
 import java.util.Map;
 
@@ -35,9 +38,12 @@ import java.util.Map;
 @Since("1.0")
 public class ExprRecursiveSize extends SimpleExpression<Long> {
 
-	static {
-		Skript.registerExpression(ExprRecursiveSize.class, Long.class, ExpressionType.PROPERTY,
-				"[the] recursive (amount|number|size) of %objects%");
+	public static void register(SyntaxRegistry syntaxRegistry) {
+		syntaxRegistry.register(SyntaxRegistry.EXPRESSION, SyntaxInfo.Expression.builder(ExprRecursiveSize.class, Long.class)
+			.supplier(ExprRecursiveSize::new)
+			.priority(PropertyExpression.DEFAULT_PRIORITY)
+			.addPattern("[the] recursive (amount|number|size) of %objects%")
+			.build());
 	}
 
 	private ExpressionList<?> exprs;
