+++ src/main/java/org/skriptlang/skript/common/properties/PropertiesModule.java
@@ -2,15 +2,19 @@ package org.skriptlang.skript.common.properties;
 
 import ch.njol.skript.SkriptConfig;
 import org.skriptlang.skript.addon.AddonModule;
+import org.skriptlang.skript.addon.HierarchicalAddonModule;
 import org.skriptlang.skript.addon.SkriptAddon;
-import org.skriptlang.skript.common.properties.conditions.PropCondContains;
-import org.skriptlang.skript.common.properties.conditions.PropCondIsEmpty;
-import org.skriptlang.skript.common.properties.expressions.*;
+import org.skriptlang.skript.common.properties.elements.conditions.*;
+import org.skriptlang.skript.common.properties.elements.expressions.*;
 
-public class PropertiesModule implements AddonModule {
+public class PropertiesModule extends HierarchicalAddonModule {
+
+	public PropertiesModule(AddonModule parentModule) {
+		super(parentModule);
+	}
 
 	@Override
-	public void load(SkriptAddon addon) {
+	protected void loadSelf(SkriptAddon addon) {
 		register(addon, PropExprScale::register);
 		if (SkriptConfig.useTypeProperties.value()) { // not using canLoad since this should only gate old properties, not new ones
 			register(addon,
