+++ src/main/java/org/skriptlang/skript/common/properties/elements/conditions/PropCondIsEmpty.java
@@ -1,4 +1,4 @@
-package org.skriptlang.skript.common.properties.conditions;
+package org.skriptlang.skript.common.properties.elements.conditions;
 
 import ch.njol.skript.conditions.base.PropertyCondition;
 import ch.njol.skript.conditions.base.PropertyCondition.PropertyType;
