+++ src/main/java/org/skriptlang/skript/common/properties/elements/expressions/PropExprAmount.java
@@ -1,6 +1,6 @@
-package org.skriptlang.skript.common.properties.expressions;
+package org.skriptlang.skript.common.properties.elements.expressions;
 
-import ch.njol.skript.classes.Changer;
+import ch.njol.skript.classes.Changer.ChangeMode;
 import ch.njol.skript.doc.*;
 import ch.njol.skript.expressions.base.PropertyExpression;
 import ch.njol.skript.lang.Expression;
@@ -9,6 +9,7 @@ import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.util.LiteralUtils;
 import ch.njol.util.Kleenean;
 import org.bukkit.event.Event;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.skriptlang.skript.lang.properties.Property;
@@ -16,21 +17,23 @@ import org.skriptlang.skript.lang.properties.PropertyBaseExpression;
 import org.skriptlang.skript.lang.properties.handlers.base.ExpressionPropertyHandler;
 import org.skriptlang.skript.registration.SyntaxRegistry;
 
-@Name("Size")
+@Name("Amount")
 @Description("""
-	The size of something.
-	Using 'size of {list::*}' will return the length of the list, so if you want the sizes of the things inside the \
-	lists, use 'sizes of {list::*}'.
+	The amount of something.
+	Using 'amount of {list::*}' will return the length of the list, so if you want the amounts of the things inside the \
+	lists, use 'amounts of {list::*}'.
 	""")
-@Example("message \"There are %size of all players% players online!\"")
-@Since({"1.0", "2.13 (sizes of)"})
-@RelatedProperty("size")
-public class PropExprSize extends PropertyBaseExpression<ExpressionPropertyHandler<?, ?>> {
+@Example("message \"There are %amount of all players% players online!\"")
+@Example("if amount of player's tool > 5:")
+@Example("if amounts of player's tool and player's offhand tool > 5:")
+@Since({"1.0", "2.13 (amounts of)"})
+@RelatedProperty("amount")
+public class PropExprAmount extends PropertyBaseExpression<ExpressionPropertyHandler<?, ?>> {
 
 	public static void register(SyntaxRegistry registry) {
 		registry.register(SyntaxRegistry.EXPRESSION,
-			PropertyExpression.infoBuilder(PropExprSize.class, Object.class, "size[:s]", "objects", false)
-				.supplier(PropExprSize::new)
+			PropertyExpression.infoBuilder(PropExprAmount.class, Object.class, "amount[:s]", "objects", false)
+				.supplier(PropExprAmount::new)
 				.build());
 	}
 
@@ -39,17 +42,34 @@ public class PropExprSize extends PropertyBaseExpression<ExpressionPropertyHandl
 
 	@Override
 	public boolean init(Expression<?>[] expressions, int matchedPattern, Kleenean isDelayed, ParseResult parseResult) {
-		// size[s] of x -> property
-		// sizes of x, y -> property
-		// size of x, y -> list length
+		// amount[s] of x -> property
+		// amounts of x, y -> property
+		// amount of x, y -> list length
 		useProperties = parseResult.hasTag("s") || expressions[0].isSingle();
 		if (useProperties) {
 			return super.init(expressions, matchedPattern, isDelayed, parseResult);
 		} else {
 			// if exprlist or varlist, count elements
-			this.exprs = PropExprAmount.asExprList(expressions[0]);
+			this.exprs = asExprList(expressions[0]);
 			return LiteralUtils.canInitSafely(this.exprs);
 		}
+	}
+
+	/**
+	 * Wraps non-expressionlists in a expression list.
+	 * @param expr The expression to wrap
+	 * @return An ExpressionList containing the original expression, or the original expression if it was already a list.
+	 *   		Null if the expression could not be converted to a valid expression list.
+	 */
+	@ApiStatus.Internal
+	public static ExpressionList<?> asExprList(Expression<?> expr) {
+		ExpressionList<?> exprs;
+		if (expr instanceof ExpressionList<?> exprList) {
+			exprs = exprList;
+		} else {
+			exprs = new ExpressionList<>(new Expression<?>[]{ expr }, Object.class, false);
+		}
+		return (ExpressionList<?>) LiteralUtils.defendExpression(exprs);
 	}
 
 	@Override
@@ -60,7 +80,7 @@ public class PropExprSize extends PropertyBaseExpression<ExpressionPropertyHandl
 	}
 
 	@Override
-	public Class<?> @Nullable [] acceptChange(Changer.ChangeMode mode) {
+	public Class<?> @Nullable [] acceptChange(ChangeMode mode) {
 		if (useProperties)
 			return super.acceptChange(mode);
 		return null;
@@ -68,7 +88,7 @@ public class PropExprSize extends PropertyBaseExpression<ExpressionPropertyHandl
 
 	@Override
 	public @NotNull Property<ExpressionPropertyHandler<?, ?>> getProperty() {
-		return Property.SIZE;
+		return Property.AMOUNT;
 	}
 
 	@Override
@@ -96,7 +116,7 @@ public class PropExprSize extends PropertyBaseExpression<ExpressionPropertyHandl
 	public String toString(Event event, boolean debug) {
 		if (useProperties)
 			return super.toString(event, debug);
-		return "size of " + this.exprs.toString(event, debug);
+		return "amount of " + this.exprs.toString(event, debug);
 	}
 
 }
