+++ src/main/java/org/skriptlang/skript/lang/properties/Property.java
@@ -12,8 +12,8 @@ import org.skriptlang.skript.addon.SkriptAddon;
 import org.skriptlang.skript.bukkit.base.types.InventoryClassInfo;
 import org.skriptlang.skript.bukkit.base.types.ItemStackClassInfo;
 import org.skriptlang.skript.bukkit.base.types.PlayerClassInfo;
-import org.skriptlang.skript.common.properties.conditions.PropCondContains;
-import org.skriptlang.skript.common.properties.expressions.PropExprName;
+import org.skriptlang.skript.common.properties.elements.conditions.PropCondContains;
+import org.skriptlang.skript.common.properties.elements.expressions.PropExprName;
 import org.skriptlang.skript.common.types.QueueClassInfo;
 import org.skriptlang.skript.common.types.ScriptClassInfo;
 import org.skriptlang.skript.lang.properties.handlers.ContainsHandler;
