+++ src/test/java/ch/njol/skript/patterns/PatternConflictsTest.java
@@ -16,7 +16,7 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.skriptlang.skript.bukkit.text.elements.ExprColored;
 import org.skriptlang.skript.bukkit.potion.elements.conditions.CondHasPotion;
-import org.skriptlang.skript.common.properties.conditions.PropCondContains;
+import org.skriptlang.skript.common.properties.elements.conditions.PropCondContains;
 import org.skriptlang.skript.lang.structure.Structure;
 import org.skriptlang.skript.registration.SyntaxInfo;
